Singular button for exporting both applications and comments data.
