+++ src/Components/AllCandidates/CommitteeReview.jsx
@@ -7,8 +7,6 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faAngleLeft } from "@fortawesome/free-solid-svg-icons";
 import { connect } from "react-redux";
 import Spinner from "react-spinner-material";
-import Button from "@material-ui/core/Button";
-import { CSVLink } from 'react-csv';
 
 const GET = require("../../requests/get");
 
@@ -139,8 +137,7 @@ class CommitteeReview extends Component {
       });
 
       this.setState({
-        committeeSize: users.length,
-        users: users
+        committeeSize: users.length
       });
       for (let i = 0; i < users.length; i++) {
         GET.getReviewCountAPI(users[i].userId).then((count) => {
@@ -150,73 +147,11 @@ class CommitteeReview extends Component {
         });
       }
     });
-    GET.getAllReviewsAPI().then(reviews => {
-      this.setState({ reviews: reviews });
-    })
-    GET.getAllApplicationsAPI().then(applications => {
-      this.setState({ applications: applications });
-    })
   }
 
   goBack() {
     this.props.history.push("/admin/allcandidates");
   }
-
-  //Calculate the average ranking
-  getComments = () => {
-    let commentsTotal = []
-    this.state.applications.sort().forEach((application, i) => {
-      let comments = []
-      this.state.reviews.forEach((review) => {
-        if (
-          review.applicationId === application._id &&
-          review.userId !== "vBUgTex5MeNd57fdB8u4wv7kXZ52" &&
-          review.userId !== "hM9QRmlybTdaQkLX25FupXqjiuF2"
-        ) {
-          //Go through questions and tally the comments
-          let temp = []
-          review.comments.forEach(comment =>{
-            temp.push(comment)
-          });
-          review.questionList.forEach(question =>{
-            question.notes.forEach(note => {
-              temp.push(note)
-            });
-          });
-          let email = ""
-          this.state.users.forEach(user => {
-            if (user.userId === review.userId){
-              email = user.email
-            }
-          });
-
-          let comment = {
-            comments: temp,
-            userId: email
-          }
-          if (comment.comments.length !== 0){
-            comments.push(comment)
-          }
-
-        }
-      })
-      let newComment = {
-        Name: application["Organization Name"] + ' (Total Commenters: ' + comments.length + ')',
-        Comments: ""
-      }
-      commentsTotal.push(newComment)
-      comments.forEach(comment => {
-        comment.comments.forEach(c => {
-          let newComment = {
-            Name: "",
-            Comments: c.value + ' (' + comment.userId + ')'
-          }
-          commentsTotal.push(newComment)
-        });
-      });
-    });
-    this.setState({ comments: commentsTotal });
-  };
 
   render() {
     return (
@@ -238,20 +173,6 @@ class CommitteeReview extends Component {
                   Back to Candidate Submissions
                 </span>
               </p>
-              <div className="button-container">
-                <CSVLink data={this.state.comments}>
-                  <Button
-                    variant="contained"
-                    color="primary"
-                    target="_blank"
-                    value="ExportData"
-                    style={{width: '250px', maxWidth: '250px'}}
-                    onClick={() => { this.getComments() }}
-                  >
-                    Export Comments
-                  </Button>
-                </CSVLink>
-              </div>
             </Header>
             <h1 align="left" style={{color: 'black'}}>Committee Review Completion</h1>
             <CommitteeReviewTable
